release: refatorando o thema
