--- src/themes/dark.ts
@@ -1,13 +1,0 @@
-import { createTheme } from '@mui/material'
-
-
-export const DarkTheme = createTheme({
-    palette: {
-        primary: {
-            main: '#f2a30b',
-        },
-        background: {
-            default: '#0b0b0b',
-        },
-    }
-})
