--- src/themes/light.ts
@@ -1,12 +1,0 @@
-import { createTheme } from '@mui/material'
-
-export const LightTheme = createTheme({
-    palette: {
-        primary: {
-            main: '#FF4500',
-        },
-        background: {
-            default: '#f2a30b',
-        }
-    }
-})
